+++ R/psut.R
@@ -9,6 +9,8 @@
 #' `.tidy_iea_df` is typically obtained from `tidy_iea_df()`.
 #'
 #' @param .tidy_iea_df the tidy data frame from which a unit summation `S_units` matrix is to be formed.
+#' @param class The type of matrix to be created, one of "matrix" or "Matrix".
+#'              Default is "matrix".
 #' @param ledger_side,flow_aggregation_point,flow,product,e_dot,unit,matnames See `IEATools::iea_cols`.
 #' @param s_units See `IEATools::psut_cols`.
 #' @param product_type,unit_type See `IEATools::row_col_types`.
@@ -25,6 +27,7 @@
 #' load_tidy_iea_df() %>% 
 #'   extract_S_units_from_tidy()
 extract_S_units_from_tidy <- function(.tidy_iea_df, 
+                                      class = c("matrix", "Matrix"),
                                       # Column names in .tidy_iea_df
                                       ledger_side = IEATools::iea_cols$ledger_side, 
                                       flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point, 
@@ -42,6 +45,8 @@ extract_S_units_from_tidy <- function(.tidy_iea_df,
                                       .val = ".val", 
                                       .rowtype = ".rowtype", 
                                       .coltype = ".coltype"){
+  
+  class <- match.arg(class)
   
   # grouping_vars <- matsindf::everything_except(.tidy_iea_df, ledger_side, flow_aggregation_point, flow, product, e_dot, unit, matnames)
   grouping_symbols <- matsindf::everything_except(.tidy_iea_df, ledger_side, flow_aggregation_point, flow, product, e_dot, unit, matnames)
@@ -63,7 +68,8 @@ extract_S_units_from_tidy <- function(.tidy_iea_df,
     ) %>%
     matsindf::collapse_to_matrices(matnames = s_units, matvals = .val,
                                    rownames = product, colnames = unit,
-                                   rowtypes = .rowtype, coltypes = .coltype) %>%
+                                   rowtypes = .rowtype, coltypes = .coltype, 
+                                   class = class) %>%
     dplyr::rename(
       # "{s_units}" := .data[[.val]]
       "{s_units}" := dplyr::all_of(.val)
@@ -306,6 +312,8 @@ add_row_col_meta <- function(.tidy_iea_df,
 #' `dplyr::ungroup()` prior to undergoing any modification.
 #'
 #' @param .tidy_iea_df a data frame containing `matnames` and several other columns
+#' @param class The type of matrix to be created, one of "matrix" or "Matrix".
+#'              Default is "matrix".
 #' @param ledger_side,flow_aggregation_point,flow,product,e_dot,unit See `IEATools::iea_cols`.
 #' @param matnames,rownames,colnames,rowtypes,coltypes See `IEATools::mat_meta_cols`.
 #' @param matvals See `IEATools::psut_cols`.
@@ -322,6 +330,7 @@ add_row_col_meta <- function(.tidy_iea_df,
 #'   add_row_col_meta() %>% 
 #'   collapse_to_tidy_psut()
 collapse_to_tidy_psut <- function(.tidy_iea_df,
+                                  class = c("matrix", "Matrix"),
                                   # Names of input columns
                                   ledger_side = IEATools::iea_cols$ledger_side,
                                   flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point, 
@@ -337,6 +346,9 @@ collapse_to_tidy_psut <- function(.tidy_iea_df,
                                   coltypes = IEATools::mat_meta_cols$coltypes, 
                                   # Name of output column of matrices
                                   matvals = IEATools::psut_cols$matvals){
+  
+  class <- match.arg(class)
+  
   matsindf::verify_cols_missing(.tidy_iea_df, matvals)
   
   .tidy_iea_df %>% 
@@ -364,7 +376,8 @@ collapse_to_tidy_psut <- function(.tidy_iea_df,
     # Now we can collapse!
     matsindf::collapse_to_matrices(matnames = matnames, matvals = e_dot,
                                    rownames = rownames, colnames = colnames,
-                                   rowtypes = rowtypes, coltypes = coltypes) %>%
+                                   rowtypes = rowtypes, coltypes = coltypes, 
+                                   class = class) %>%
     dplyr::rename(
       # "{matvals}" := .data[[e_dot]]
       "{matvals}" := dplyr::all_of(e_dot)
@@ -557,6 +570,8 @@ replace_null_RUV <- function(.sutmats = NULL,
 #' the return value is a zero-row data frame with expected columns.
 #'
 #' @param .tidy_iea_df a tidy data frame that has been specified with `specify_all()`.
+#' @param class The type of matrix to be created, one of "matrix" or "Matrix".
+#'              Default is "matrix".
 #' @param year,ledger_side,flow_aggregation_point,flow,product,e_dot,unit See `IEATools::iea_cols`.
 #' @param supply,consumption See `IEATools::ledger_sides`.
 #' @param matnames,rownames,colnames,rowtypes,coltypes See `IEATools::mat_meta_cols`.
@@ -601,6 +616,7 @@ replace_null_RUV <- function(.sutmats = NULL,
 #'   as.logical() %>% 
 #'   all()
 prep_psut <- function(.tidy_iea_df, 
+                      class = c("matrix", "Matrix"),
                       year = IEATools::iea_cols$year,
                       ledger_side = IEATools::iea_cols$ledger_side, 
                       flow_aggregation_point = IEATools::iea_cols$flow_aggregation_point,
@@ -625,6 +641,9 @@ prep_psut <- function(.tidy_iea_df,
                       Y = IEATools::psut_cols$Y,
                       B = IEATools::psut_cols$B,
                       s_units = IEATools::psut_cols$s_units){
+  
+  class <- match.arg(class)
+  
   if (nrow(.tidy_iea_df) == 0) {
     # We can get a no-row data frame for .tidy_iea_df. 
     # If so, we should return a no-row data frame with empty columns added.
@@ -648,6 +667,7 @@ prep_psut <- function(.tidy_iea_df,
   
   # We actually have some rows in .tidy_iea_df, so work with them
   S_units <- extract_S_units_from_tidy(.tidy_iea_df, 
+                                       class = class,
                                        product = product, 
                                        unit = unit)
   # Bundle functions together
@@ -657,7 +677,8 @@ prep_psut <- function(.tidy_iea_df,
     # Add additional metadata
     add_row_col_meta(flow = flow, product = product, matnames = matnames) %>% 
     # Now collapse to matrices
-    collapse_to_tidy_psut(e_dot = e_dot, matnames = matnames, matvals = matvals, rownames = rownames, colnames = colnames,
+    collapse_to_tidy_psut(class = class, e_dot = e_dot, matnames = matnames, matvals = matvals, 
+                          rownames = rownames, colnames = colnames,
                           rowtypes = rowtypes, coltypes = coltypes) 
   # Get a list of matrix names for future use
   matrix_names <- Collapsed[[matnames]] %>%